New profile page, and like button
Have made a profile page so that users can favourite products on the comparison page and see them on their profile. Button is not functional yet. Also implemented Emails for the contact page so user can send emails. Adjusted about us page too. Nav bar is dynamic now as well if user is logged in or not.